+++ js/src/forum/index.ts
@@ -54,7 +54,11 @@ function addSourceLinkReplacement() {
       if (match == null) {
         return;
       }
-      if (a.text === a.href && !a.classList.contains('DiscussionLink')) {
+      if (
+        app.forum.attribute('retrofitLinksInFrontend')
+        && a.text === a.href
+        && !a.classList.contains('DiscussionLink')
+      ) {
         const discussionId = match[1];
         const span = document.createElement('span');
         m.mount(span, {view: () => m(DiscussionLink, {discussionId, href: a.href})});